move monitored namespaces to shared package, remove dynamic fetch
Move the scoped namespace list to pkg/util/namespace/monitored.go as a
shared exported variable. Initialize namespacesToMonitor directly in the
Monitor constructor instead of a separate fetch step, eliminating the
now-unnecessary managednamespaces.go and its redundant test file.
